// src/app/models/enums/kiosk-action.enum.ts
export enum KioskAction {
  DeviceCode = 'DEVICE_CODE',
  RedirectToKioskExecution = 'REDIRECT_TO_KIOSK_EXECUTION',
  RefreshKioskDevice = 'REFRESH_KIOSK_DEVICE',
  standByKioskDevice = 'STANDBY_KIOSK_DEVICE',
  deRegisterKioskDevice = 'DE_REGISTER_KIOSK_DEVICE',
  sendMessageKioskDevice = 'SEND_MESSAGE_KIOSK_DEVICE',
  shutDownKioskDevice = 'SHUTDOWN_KIOSK_DEVICE',
  addedInWaiting = 'ADDED_IN_WAITING',
  removedInWaiting = 'REMOVED_IN_WAITING',
  nowCallingAdded = 'NOW_CALLING_ADDED',
  nowCallingRemoved = 'NOW_CALLING_REMOVED',
  waitingAdded = 'WAITING_ADDED',
  waitingRemoved = 'WAITING_REMOVED',
  nowServingAdded = 'NOW_SERVING_ADDED',
  nowServingRemoved = 'NOW_SERVING_REMOVED',
  updateMonitor = 'MONITOR_UPDATE',
  updateMobileMonitor = 'MOBILE_MONITOR_UPDATE',
}